extract spec parsing into a separate package
